+++ static/css/main.css
@@ -133,7 +133,6 @@ a {
 
 .hint__select-date {
   color: red;
-  display: inline-block;
   padding-left: 1rem;
   display: none;
 }
