+++ static/js/attraction.js
@@ -184,3 +184,16 @@ selecTimeRadioBtns.forEach((btn) => {
 });
 
 render(indexOfAttractionIdInHref);
+
+function generateDate() {
+  const today = new Date();
+  const year = today.getFullYear();
+  const month = String(today.getMonth() + 1).padStart(2, "0");
+  const day = String(today.getDate()).padStart(2, "0");
+
+  const formattedDate = `${year}-${month}-${day}`;
+  const dateInput = document.querySelector("#date");
+  dateInput.setAttribute("min", formattedDate);
+}
+
+generateDate();
